code/requirements.txt
opencv-python
torchvision
ultralytics
lap